# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Approval Flow API",
    "description": "Public REST API for Smart Approval Flow. All responses use a consistent envelope. See docs/api-design.md.",
    "version": "1.0.0-block1"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session JWT (users) or API key (integrations, Block 5)."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/templates": {
      "get": {
        "summary": "List message templates for the current tenant",
        "tags": [
          "templates"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/templates/{id}": {
      "get": {
        "summary": "Read one message template",
        "tags": [
          "templates"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "patch": {
        "summary": "Update a template’s subject and/or body",
        "tags": [
          "templates"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/public/approvals/{token}": {
      "get": {
        "summary": "Read an approval request via its secure token (no login)",
        "tags": [
          "public"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/public/approvals/{token}/respond": {
      "post": {
        "summary": "Submit a customer decision (approve / decline / callback)",
        "tags": [
          "public"
        ],
        "description": "Loginless. Idempotent via the optional Idempotency-Key header so a double-tap on mobile does not create two decisions.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
}